requirements.txt
Adafruit-Blinka==8.66.2
adafruit-circuitpython-busdevice==5.2.14
adafruit-circuitpython-connectionmanager==3.1.6
adafruit-circuitpython-dht==4.0.10
adafruit-circuitpython-requests==4.1.14
adafruit-circuitpython-sht31d==2.3.29
adafruit-circuitpython-typing==1.12.2
Adafruit-PlatformDetect==3.84.1
Adafruit-PureIO==1.1.11
binho-host-adapter==0.1.6
blinker==1.9.0
click==8.3.0
colorzero==2.0
Flask==3.1.2
flask-cors==6.0.0
gpiozero==2.0.1
itsdangerous==2.2.0
Jinja2==3.1.6
MarkupSafe==3.0.3
pyftdi==0.57.1
pyserial==3.5
pyusb==1.3.1
RPi.GPIO==0.7.1
RPLCD==1.4.0
setuptools==80.9.0
smbus2==0.5.0
sysv_ipc==1.1.0
typing_extensions==4.15.0
Werkzeug==3.1.3
